pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-autonomous-assistant"
version = "0.1.0"
description = "AI Autonomous Assistant"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE.txt"]
keywords = ["AI", "Autonomous", "Assistant","Raspberry PI"]
authors = [
    {name = "Chcavignx", email = "chcavignx@gmail.com"}
]
maintainers = [
    {name = "Chcavignx", email = "chcavignx@gmail.com"}
]

classifiers = [
  # How mature is this project? Common values are
  #   1 - Planning
  #   2 - Pre-Alpha
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 2 - Pre-Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",

  # Specify the Python versions you support here.
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

# Dépendances principales (runtime)
dependencies = [
    "blobfile",
    "flask-cors>=4.0.0",
    "flask>=2.3.0",
    "joblib>=1.3.0",
    "librosa>=0.10.0",
    "loguru>=0.7.0",
    "numba>=0.57.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "python-dotenv",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "types-requests",
    "typing-extensions>=4.12",
]

# Dépendances optionnelles pour CI/CD
[project.optional-dependencies]
ai = [
    "accelerate>=0.20.3",
    "datasets>=2.14.5",
    "huggingface-hub>=0.18.0",
    "tiktoken>=0.4.0",
    "tokenizers>=0.19.0",
    "transformers>=4.35.0",
]
vision = [
    "pillow>=10.0.0",
    "torch>=2.0.2",
    "torchvision>=0.15.2",
]
audio = [
    "faster-whisper>=0.1.3",
    "openai-whisper>=20230314",
    "piper-tts>=0.1.0",
    "pydub>=0.25.1",
    "pyttsx3>=2.90",
    "scipy>=1.11.0",
    "silero-vad>=2.0.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "speechrecognition>=3.10.0",
    "torchaudio>=2.0.2",
    "vosk>=0.3.45",
]
dev = [
    "basedpyright>=1.31.7",
    "gitingest",
    "mypy>=1.5.0",
    "poethepoet",
    "pre-commit",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.0.0",
    "radon",
    "ruff",
    "skylos",
]
rpi = [
    "pyaudio>=0.2.14",
]
all = [
    "ai-autonomous-assistant[audio,ai,vision,dev,rpi]",
]

[tool.uv]
# Configuration spécifique à uv
# uv will use the python_version from [project] requires-python
override-dependencies = [
    "numpy>=1.26.0,<2.0"
]
# Support multiple platforms: macOS (dev), Linux x86_64 (CI), Linux ARM64 (CI/RPI)
environments = [
    "sys_platform == 'darwin' and platform_machine == 'arm64'",  # macOS (local development)
    "sys_platform == 'linux' and platform_machine == 'aarch64'"  # ARM64 (RPI/CI)
]

[tool.mypy]
python_version = "3.11"
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
check_untyped_defs = true
namespace_packages = true
explicit_package_bases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py", "__test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-W error",  # STRICT: Treat warnings as errors (catch deprecations early)
    "--tb=short",
    "--cov=.",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",  # STRICT: Require 80% coverage
]
filterwarnings = [
    "error",  # Treat all warnings as errors by default
    # Uncomment to selectively ignore specific third-party warnings if needed:
    # "ignore::DeprecationWarning:some_noisy_library",
    # "ignore::UserWarning:another_library",
]
markers = [
    "basic: marks basic infrastructure tests",
    "integration: marks integration or long-running tests",
]

norecursedirs = [
"**/__pycache__",
"**/tests/__pycache__",
"**/*.pyc",
"**/dist/",
"**/*.egg-info/",
"**/.pytest_cache/",
"**/.mypy_cache/",
"**/.ruff_cache/",
"**/examples/",
"**/test*",
"**/node_modules",
".git",
".pyright_cache",
".venv",
"venv",
"env",
"logs",
"outputs",
"*.egg-info",
]
